+++ src/main/java/com/seoultech/ecc/member/repository/LevelChangeRequestRepository.java
@@ -7,8 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface LevelChangeRequestRepository extends JpaRepository<LevelChangeRequestEntity, Long> {
+public interface LevelChangeRequestRepository extends JpaRepository<LevelChangeRequestEntity, Integer> { // Long -> Integer 변경
 
+    // @Query 제거하고 JPA 네이밍 컨벤션 적용
     List<LevelChangeRequestEntity> findByStatus(LevelChangeRequestEntity.RequestStatus status);
 
     List<LevelChangeRequestEntity> findByMember(MemberEntity member);
