+++ src/main/java/com/seoultech/ecc/member/repository/MajorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MajorRepository extends JpaRepository<MajorEntity, Long> {
+public interface MajorRepository extends JpaRepository<MajorEntity, Integer> { // Long -> Integer 변경
     Optional<MajorEntity> findByName(String name);
 }
